remove course extractor trait
