+++ crates/extractor/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 [dependencies]
 anyhow = "1.0.86"
 model = { path = "../../crates/model" }
-scraper = "0.16.0"
+scraper = "0.23.1"
 
 [dev-dependencies]
 include_dir = "0.7.4"
